feat: passing VM via context API
