+++ src/components/dashboard/Position/PositionExpanded/index.tsx
@@ -7,7 +7,7 @@ import { RowBetween, RowFixed } from 'components/Row'
 import { MouseoverTooltip } from 'components/Tooltip'
 import { Snapshot } from 'data/dashboard/expandedData'
 import useTheme from 'hooks/useTheme'
-import React from 'react'
+import React, { useContext } from 'react'
 import { useExpandedData } from 'state/dashboard/hooks'
 import { usePoolDatas } from 'state/pools/hooks'
 import styled from 'styled-components'
@@ -18,6 +18,7 @@ import Icon from 'components/Icon'
 import dayjs from 'dayjs'
 import utc from 'dayjs/plugin/utc'
 import Interactions from './Interactions'
+import { VMContext } from 'pages/App'
 
 dayjs.extend(utc)
 
@@ -179,7 +180,6 @@ interface Props {
   token1priceUSD: number
   token0CurrentAmount: number
   token1CurrentAmount: number
-  vm: any
 }
 
 const PositionExpanded = ({
@@ -196,9 +196,9 @@ const PositionExpanded = ({
   token1priceUSD,
   token0CurrentAmount,
   token1CurrentAmount,
-  vm,
 }: Props) => {
   const theme = useTheme()
+  const vm = useContext(VMContext)
   const expandedInfo = useExpandedData(owner, tokenId, vm)
   // liquidity distribution data
   const poolData = usePoolDatas([poolAddress])[0]
